fixed join and leave commands' handlers

+++ src/main/java/gp/assessments/chat/common/entity/ChatChannelEntity.java
@@ -11,9 +11,9 @@ public class ChatChannelEntity {
 
     private final String id;
     private String channelName;
-    private long channelLimit;
+    private int channelLimit;
 
-    public ChatChannelEntity(String channelName, long channelLimit) {
+    public ChatChannelEntity(String channelName, int channelLimit) {
         this.id = UUID.randomUUID().toString();
         this.channelName = channelName;
         this.channelLimit = channelLimit;
